Fix: 有一定概率不会scroll to top

+++ CCScrollableHeaderTabViewController/EmbedViewController.m
@@ -42,7 +42,7 @@
         cell = [[UITableViewCell alloc] initWithStyle:UITableViewCellStyleDefault reuseIdentifier:kCellId];
     }
     
-    cell.textLabel.text = [NSString stringWithFormat:@"hello %d", indexPath.row];
+    cell.textLabel.text = [NSString stringWithFormat:@"hello %ld", (long)indexPath.row];
     return cell;
 }
 @end
